a minor improvement on pss trojan mining

- code cleanup
- using errgroup
- change to bigendian 
- fix for the parity bit

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/ethersphere/bee/2140)
<!-- Reviewable:end -->