@acud 2021-06-21
before:
```
growlie@growlie-strikes-back ~/go/src/github.com/ethersphere/bee/pkg/pss ±master⚡ » go test -v -bench=. -run Bench -benchmem -benchtime 5s
goos: linux
goarch: amd64
pkg: github.com/ethersphere/bee/pkg/pss
cpu: Intel(R) Core(TM) i7-7500U CPU @ 2.70GHz
BenchmarkWrap
BenchmarkWrap/length:1,depth:1
BenchmarkWrap/length:1,depth:1-4         	     187	  33675828 ns/op	18788199 B/op	  104356 allocs/op
BenchmarkWrap/length:4,depth:1
BenchmarkWrap/length:4,depth:1-4         	     622	   9896097 ns/op	 5419940 B/op	   30071 allocs/op
BenchmarkWrap/length:16,depth:1
BenchmarkWrap/length:16,depth:1-4        	    1520	   3850104 ns/op	 2013067 B/op	   11114 allocs/op
BenchmarkWrap/length:16,depth:2
BenchmarkWrap/length:16,depth:2-4        	      30	 757000467 ns/op	425943036 B/op	 2366465 allocs/op
BenchmarkWrap/length:64,depth:2
BenchmarkWrap/length:64,depth:2-4        	      38	 145133468 ns/op	81234541 B/op	  451305 allocs/op
BenchmarkWrap/length:256,depth:2
BenchmarkWrap/length:256,depth:2-4       	     166	  37935671 ns/op	21093620 B/op	  117162 allocs/op
BenchmarkWrap/length:256,depth:3
BenchmarkWrap/length:256,depth:3-4       	       1	6143139754 ns/op	3259615688 B/op	18111114 allocs/op
BenchmarkWrap/length:4096,depth:3
BenchmarkWrap/length:4096,depth:3-4      	      15	 591351344 ns/op	242435498 B/op	 1347379 allocs/op
BenchmarkWrap/length:16384,depth:3
BenchmarkWrap/length:16384,depth:3-4     	      40	 223228374 ns/op	74415670 B/op	  413806 allocs/op
PASS
ok  	github.com/ethersphere/bee/pkg/pss	106.122s
```

after:
```
growlie@growlie-strikes-back ~/go/src/github.com/ethersphere/bee/pkg/pss ±(dfccf46b...)⚡ » go test -v -bench=. -run Bench -benchmem -benchtime 5s
goos: linux
goarch: amd64
pkg: github.com/ethersphere/bee/pkg/pss
cpu: Intel(R) Core(TM) i7-7500U CPU @ 2.70GHz
BenchmarkWrap
BenchmarkWrap/length:1,depth:1
BenchmarkWrap/length:1,depth:1-4         	     232	  27393762 ns/op	19365715 B/op	  107572 allocs/op
BenchmarkWrap/length:4,depth:1
BenchmarkWrap/length:4,depth:1-4         	     844	   7616651 ns/op	 5263813 B/op	   29211 allocs/op
BenchmarkWrap/length:16,depth:1
BenchmarkWrap/length:16,depth:1-4        	    1869	   3619230 ns/op	 1997974 B/op	   11042 allocs/op
BenchmarkWrap/length:16,depth:2
BenchmarkWrap/length:16,depth:2-4        	      13	 435784661 ns/op	261811824 B/op	 1454861 allocs/op
BenchmarkWrap/length:64,depth:2
BenchmarkWrap/length:64,depth:2-4        	      61	 113410167 ns/op	75256698 B/op	  418116 allocs/op
BenchmarkWrap/length:256,depth:2
BenchmarkWrap/length:256,depth:2-4       	     206	  30770748 ns/op	18623767 B/op	  103452 allocs/op
BenchmarkWrap/length:256,depth:3
BenchmarkWrap/length:256,depth:3-4       	       1	9820205532 ns/op	3980843456 B/op	22129769 allocs/op
BenchmarkWrap/length:4096,depth:3
BenchmarkWrap/length:4096,depth:3-4      	       9	 956932108 ns/op	326600746 B/op	 1815988 allocs/op
BenchmarkWrap/length:16384,depth:3
BenchmarkWrap/length:16384,depth:3-4     	      26	 264735541 ns/op	85486048 B/op	  475630 allocs/op
PASS
ok  	github.com/ethersphere/bee/pkg/pss	80.735s
```