@aloknerurkar 2021-06-25
___
*[pkg/pss/trojan.go, line 248 at r1](https://reviewable.io/reviews/ethersphere/bee/2140#-McYRWnsCVYTPvnuwnKi-r1-248:-Md1fe9zAgxeGPVrZQFP:b5dpars) ([raw file](https://github.com/ethersphere/bee/blob/dfccf46bdaed159ef114f7da53fcbcd9964a3cef/pkg/pss/trojan.go#L248)):*
<details><summary><i>Previously, zelig (Viktor Trón) wrote…</i></summary><blockquote>

this was only to trigger the context cancel to terminate the other routines, but i guess i should just use a cancel on the context right?
</blockquote></details>

Context cancellation will return error on eg.Wait from the next routine. If we dont care about the errors from the other routines once the first one is done, we should `return nil` on `<-ctx.Done()`. Although I dont think errorgroup usage give any advantages here apart from context semantics


<!-- Sent from Reviewable.io -->