@mrekucci 2021-06-25
> _[pkg/pss/trojan.go, line 248 at r1](https://reviewable.io/reviews/ethersphere/bee/2140#-McYRWnsCVYTPvnuwnKi-r1-248:-Md1fe9zAgxeGPVrZQFP:b5dpars) ([raw file](https://github.com/ethersphere/bee/blob/dfccf46bdaed159ef114f7da53fcbcd9964a3cef/pkg/pss/trojan.go#L248)):_
> Previously, zelig (Viktor Trón) wrote…
> 
> Context cancellation will return error on eg.Wait from the next routine. If we dont care about the errors from the other routines once the first one is done, we should `return nil` on `<-ctx.Done()`. Although I dont think errorgroup usage give any advantages here apart from context semantics

Agree, we can also specifically check if the returned error is the `context.Canceled` error and ignore it in the it is.